Final_Submission/requirements.txt
ultralytics==8.0.196
opencv-python
numpy
pyyaml
streamlit
torch
torchvision
torchaudio
